moving strategyoperators package to the root folder
