+++ strategyoperators/combinators.go
