+++ strategyoperators/combinators_test.go
