+++ strategyoperators/operators.go
