+++ strategyoperators/operators_test.go
