+++ strategyoperators/strategyoperators.go
