+++ strategyoperators/strategyoperators_test.go
